+++ src/main/kotlin/com/yourssu/scouter/masterdata/support/exception/PartNotFoundException.kt
@@ -1,5 +1,6 @@
-package com.yourssu.scouter.common.support.implement.exception
+package com.yourssu.scouter.masterdata.support.exception
 
+import com.yourssu.scouter.common.support.implement.exception.CustomException
 import org.springframework.http.HttpStatus
 
 class PartNotFoundException(
