+++ src/main/kotlin/com/yourssu/scouter/recruiting/rubric/storage/InterviewRubricRepositoryImpl.kt
@@ -8,7 +8,7 @@ import com.yourssu.scouter.recruiting.rubric.implement.InterviewRubric
 import com.yourssu.scouter.recruiting.rubric.implement.InterviewRubricMapper
 import com.yourssu.scouter.recruiting.rubric.implement.InterviewRubricReader
 import com.yourssu.scouter.recruiting.rubric.implement.InterviewRubricWriter
-import com.yourssu.scouter.common.support.implement.exception.SemesterNotFoundException
+import com.yourssu.scouter.masterdata.support.exception.SemesterNotFoundException
 import com.yourssu.scouter.recruiting.support.implement.exception.InterviewRubricNotFoundException
 import org.springframework.stereotype.Repository
