+++ src/main/kotlin/com/yourssu/scouter/recruiting/schedule/business/ScheduleService.kt
@@ -16,7 +16,7 @@ import com.yourssu.scouter.recruiting.schedule.implement.ScheduleWriter
 import com.yourssu.scouter.recruiting.support.implement.exception.ApplicantNotFoundException
 import com.yourssu.scouter.recruiting.support.implement.exception.ScheduleNotFoundException
 import com.yourssu.scouter.masterdata.part.implement.PartReader
-import com.yourssu.scouter.common.support.implement.exception.PartNotFoundException
+import com.yourssu.scouter.masterdata.support.exception.PartNotFoundException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.Duration
