+++ src/test/kotlin/com/yourssu/scouter/recruiting/schedule/business/ScheduleServiceTest.kt
@@ -12,7 +12,7 @@ import com.yourssu.scouter.recruiting.support.implement.exception.ApplicantNotFo
 import com.yourssu.scouter.recruiting.support.implement.exception.DuplicateScheduleException
 import com.yourssu.scouter.masterdata.part.implement.fixture.PartFixtureBuilder
 import com.yourssu.scouter.masterdata.part.implement.PartReader
-import com.yourssu.scouter.common.support.implement.exception.PartNotFoundException
+import com.yourssu.scouter.masterdata.support.exception.PartNotFoundException
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
